+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## v0.1.9
+
 - Add employeeFaxNumber local data source attribute
 - Add roomIdentifier and contactOrganizationIdentifier to employeeWorkAddress local data source attribute
 - Prepare campusonline-api cache